requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
audioread==3.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coqpit==0.0.17
cycler==0.12.1
decorator==5.1.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt
django-rest-swagger
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
gast==0.5.4
gensim==4.3.2
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.6
inflect==7.0.0
ipython==8.19.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.5.0
keras==2.15.0
kiwisolver==1.4.5
konlpy==0.6.0
libclang==16.0.6
librosa==0.8.0
llvmlite==0.41.1
lxml==5.0.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
# mecab==0.996.3
ml-dtypes==0.2.0
mpmath==1.3.0
mysqlclient==2.2.1
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
prompt-toolkit==3.0.43
protobuf==4.23.4
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.1
pypinyin==0.50.0
pysbd==0.3.4
python-dateutil==2.8.2
python-mecab-ko==1.3.3
python-mecab-ko-dic==2.1.1.post2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.12.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.32.0.post1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor==2.4.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.7
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
xgboost==2.0.2
drf-yasg